two winners changed to one
Board originally has two winners down the middle and right side. It's changed to one winner on the right side now.
Winning combos would be either [1,4,7] or [2,5,8]. So either could be returned as the winner depending on the order the combo. I had to move [2,5,8] ahead of [1,4,7] in WIN_COMBINATIONS to get it to pass.