Remove unnecessary break and return statements
Credit to Kayla for both this and the last change by the way.
Board.turn_count always worked fine even though turn zero is returned as nil, so the tap to force returning 0 is gone.
The break statements in both Computer.win_or_block and Computer.counter_corner are gone too.
They were micro-optimizations that are more likely to cause bugs than have any noticeable impact, even when it comes to Wargames mode.