Make executable, and adjust whitespace usage
The bin/tictactoe file is now executable and there's a line of whitespace between getting user input and the if statement to process it.
Added whitespace in the Computer.corners_or_edges method so the detect loops are easier to read.
For consistency the WIN_COMBINATIONS constant now has spaces between the start of the array and the array's contained within.