Removed turn conditional from Computer.win_or_block, and more whitespace changes
To elaborate, the win_or_block method used to check which turn it was in order to prevent it from running on turns that you couldn't win on/the final turn.
This was a premature optimization that based on my testing actually slowed the game down more then it ever helped.
In the Computer and Game classes modified the variables trio and trinity so there's more spacing between the nested array's they contain, as they were a pain to read.
Likewise Game.winner had a nested array that has had space added.
Added a line of white space after the last method in the following classes: board, game, and player.